Began adding to function definitions | Changing method definitions as well starting with strike
- What I was working on: Changed function definitions in strike.luau from Strike:level{level} to Strike:getLevel{level}. This is because it would be easier to return an object that represents and functions as attack with it's own method of attacking.

For example, if I return an object that represents attacking, this object could be the code that will cause the character to animate, as well as setting variables that allow the character to do damage. I could have this object have a method called 'attack', so if I define this attack in another script like:

"
local strike = strike:getLevel1()
strike:attack()
"

this 'strike' object could also hold all the variables I need in the attackCommand script in relation to the attack itself.

- What I intend to do: Implement