Got input_check set up, disabled most code for Client.luau and Server.luau so old code wont run.
Now I need to implement the checks from my old code into the new for these inputs. Also may need to turn the input_check_EVENT to run the lock-on code for the client, and just return nothing for things not needing to be sent back to the client. Also have to worry about GameVariables, as majority of that entire file will be put on the server. Animations may need to be server-based, as NPCs animations will need to be run on the server. However, I should make player animations run on the client. Follow Miro graph.