Got lock-on working (so far), changed GameVariables to ClientVariables
Had to disable animations in ClientVariables cause I disabled them due to disabling the old modules which initialized them

Now I'm converting ServerInit stuff to character_init, or at least most of it.

Need to get animations working. If player, make them on client. If NPC, make them on server

Then work on the rest of the functionalities such as attacking, blocking, etc