Got permission to use clothing. Whens Alpha???
In fact I only use their clothing for this game so I wouldnt need to go out and ask more creators. Had to replace all my game images to make sure their clothing is being used in the props and stuff.

Now I need to make sure the game is functionally ready to be in Alpha.